Datatable incorporation on match schedule

+++ src/public/js/general.js
@@ -1,12 +1,14 @@
 const logOut = async () => {
-    await $.ajax({
-        type: 'DELETE',
-        url: '/authorize',
-        success: () => {
-            window.location.href = window.location.origin; // Send to main page.
-        },
-        onerror: (error) => {
-            toastr.error(error);
-        }
-    });
+    if (confirm("Do you really want to log out?")){
+        await $.ajax({
+            type: 'DELETE',
+            url: '/authorize',
+            success: () => {
+                window.location.href = window.location.origin; // Send to main page.
+            },
+            onerror: (error) => {
+                toastr.error(error);
+            }
+        });
+    }
 }
